Add share button and color toggle button UI components
