+++ iOS/MMIWG2S/AppDelegate.swift
@@ -19,18 +19,18 @@ import UIKit
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
-  var window: UIWindow?
+    var window: UIWindow?
 
-  func application(
-    _ application: UIApplication,
-    didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-  ) -> Bool {
+    func application(
+        _ application: UIApplication,
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
+    ) -> Bool {
 
-    window = UIWindow(frame: UIScreen.main.bounds)
-    window?.rootViewController = FacesViewController()
-    window?.makeKeyAndVisible()
+        window = UIWindow(frame: UIScreen.main.bounds)
+        window?.rootViewController = FacesViewController()
+        window?.makeKeyAndVisible()
 
-    return true
-  }
+        return true
+    }
 }
