+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
