+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-16@2x.png
