+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-20@3x.png
