+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-29@2x.png
