+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-29@3x.png
