+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-32@2x.png
