+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-40@2x.png
