+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-60@2x-1.png
