+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-60@3x-1.png
