+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-60@3x.png
