+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-76@2x.png
