+++ iOS/MMIWG2S/Resources/Assets.xcassets/AppIcon.appiconset/Icon-83.5@2x.png
