+++ iOS/MMIWG2S/Resources/Assets.xcassets/Contents.json
