+++ iOS/MMIWG2S/Resources/Face.scnassets/Settings.json
