+++ iOS/MMIWG2S/Resources/Face.scnassets/black_hand_100_texture.png
