+++ iOS/MMIWG2S/Resources/Face.scnassets/black_hand_80_texture.png
