+++ iOS/MMIWG2S/Resources/Face.scnassets/ear_fur.png
