+++ iOS/MMIWG2S/Resources/Face.scnassets/fox_face.scn
