+++ iOS/MMIWG2S/Resources/Face.scnassets/hidden_face.scn
