+++ iOS/MMIWG2S/Resources/Face.scnassets/nose_fur.png
