+++ iOS/MMIWG2S/Resources/Face.scnassets/red_hand_100_texture.png
