+++ iOS/MMIWG2S/Resources/Face.scnassets/red_hand_80_texture.png
