+++ iOS/MMIWG2S/Resources/Face.scnassets/red_hand_full_mouth.png
