+++ iOS/MMIWG2S/Resources/Localizable.strings
@@ -1,0 +1,7 @@
+/* 
+  Localizable.strings
+  MMIWG2S
+
+  Created by Callum Johnston on 8/16/20.
+  Copyright © 2020 Google LLC. All rights reserved.
+*/
